SPARK-4843 [YARN] Squash ExecutorRunnableUtil and ExecutorRunnable
ExecutorRunnableUtil is a parent of ExecutorRunnable because of the
yarn-alpha and yarn-stable split. Now that yarn-alpha is gone,
this commit squashes the unnecessary hierarchy.
